voks

+++ dataset.py
@@ -68,7 +68,7 @@ class AugumentedDataset(Dataset):
         real_np = np.array(real_img)
         fake_img = torch.Tensor(self.earser(real_np))
 
-        return torch.stack((real_img, fake_img)), torch.Tensor([1, 0])
+        return torch.stack((real_img, fake_img)), torch.Tensor([1, 0]).type(torch.LongTensor)
 
 def augument_data(batch_size):
     celeba_dataset = AugumentedDataset()
